% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahg.R
\name{ahg}
\alias{ahg}
\title{AHG}
\usage{
ahg(ancestry, component = "all", times = 1000, num = 10, prop = NULL)
}
\arguments{
\item{ancestry}{A numeric data frame of ancestry proportion (columns: ancestry component; rows: individual).}

\item{component}{The ancestry components of which the admixture topology is about to be inferred. Default: "all", which mean all the combinations of the input ancestry components will be used.}

\item{times}{Bootstrap number of admixture topology.}

\item{num}{The sampling size. This parameter will be ignored if the parameter "prop" is specified.}

\item{prop}{The number of sampled individuals.}
}
\value{
A list containing 1) possible admixture topologies with supporting counts and 2) matrices of correlation efficient between ancestry components.
}
\description{
This function infers the admixture topology of ancestry components with supporting counts.
}
