% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_merge.R
\name{ancmerge}
\alias{ancmerge}
\title{ancmerge}
\usage{
ancmerge(tar_anc_filelist, ref, K, poporder = NULL)
}
\arguments{
\item{tar_anc_filelist}{Character. Required. ancestry file names (recomended format: prefix.K.ancestry). The ancestry matrix file should be(2 + K) columns without header. The columns: 1st-ndividual ID; 2nd-Group ID. From the 3rd column, it indicates the ancestry proportion.}

\item{ref}{Character. Required. The reference ancestry matrix to be matched.}

\item{K}{Integer. Required. The number of the ancestry components.}

\item{poporder}{Character. (optional, input files) Population order list.}
}
\value{
A list containing the merged ancestry matrix, the supporting ratio of ancestry components, and the input ancestry matrices that match/do not match the reference.
}
\description{
This function draws a radiation plot to visualize the genetic distance
}
