% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_plot.R
\name{sectorplot}
\alias{sectorplot}
\title{Sector plot}
\usage{
sectorplot(Q, ind, target = NULL, poporder = NULL, ancescols = NULL,
  sorting = FALSE, rmin = 2, rmax = 3.7, tar.r = 0.6,
  tarangs = NULL, cendis = 1, amin = -265, amax = 85,
  tarang1 = 0, tarang2 = 180, arrow = FALSE, legend_mode = FALSE,
  ancesnames = NULL, prgap = 0.2, noline = FALSE,
  pop.lab.cex = NULL, pop.lab.col = "black", pop.lab.font = 1,
  tar.lab.cex = 6, tar.lab.col = "navy", arrow.col = "red",
  arrow.lwd = 2, legend.pos = "topright")
}
\arguments{
\item{Q}{A numeric data frame of ancestry proportion (columns: ancestry componenyt; rows: individual). e.g., an output ".Q" file of the softaware ADMIXTURE.}

\item{ind}{A two-column data frame (1: population; 2: individual)."}

\item{target}{Character. The target populations to be plotted as a pie chart in the center of the circle figure. The population must be included in input "ind" and "Q" data frame.}

\item{poporder}{Character. The populations to be included in the figure, also the display order of the populations in the figure.}

\item{ancescols}{The color code of each ancestry component in the figure.}

\item{sorting}{A logical value to define whether to sort the order of the populations, which will be musked if "poporder" is specified.}

\item{rmin}{The radium of the inner ring. Default is 2.}

\item{rmax}{The radium of the outer ring. Default is 3.7.}

\item{tar.r}{A numeric vector. The radius of the target pie chart. Defualt: 0.6.}

\item{tarangs}{A numeric vector. The angles of the target pie charts.}

\item{cendis}{A numeric vector. The distance from the center of a target pie chart to the center of the sectorplot. Default: 1.}

\item{amin}{The angle at which the ring is initiated. Defualt is -265.}

\item{amax}{The angle at which the ring is ended. Default is 85.}

\item{tarang1}{The start angle of the target leyout. Default is 0.}

\item{tarang2}{The end angle of the target leyout. Defualt is 360.}

\item{arrow}{Logical. Whether to draw the arrows to the target pies.}

\item{legend_mode}{Logical. Whether to draw the legend of ancestry components.}

\item{ancesnames}{Character. To specify the names of ancestry components. If not specified, would be shown as "Ancestry_1" "Ancestry_2" and so on.}

\item{prgap}{A numeric value for gap length. Default is 0.2.}

\item{noline}{Logical. Whether to remove the black lines between populations. Default is FALSE.}

\item{pop.lab.cex}{The cex of the population labels.}

\item{pop.lab.col}{The color of population labels. The length of this parameter can be 1 or equal to the number of the populations.Default: "black".}

\item{pop.lab.font}{The font of the population labels.}

\item{tar.lab.cex}{The cex of the target label.}

\item{tar.lab.col}{The color of the target label. Default: "navy".}

\item{arrow.col}{The color of the arrow(s). Default: "red".}

\item{arrow.lwd}{The line width of the arrow(s). Default: 2.}

\item{legend.pos}{The position of the legend. Can be set as "top", "topleft", "topright", "bottom", "bottomleft", "bottomright", "left", and "right".}
}
\description{
This function draws a sector plot to visualize ancestry proportion.
}
