% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_plot.R
\name{radiationplot}
\alias{radiationplot}
\title{Radiation plot}
\usage{
radiationplot(data, target = "target", cenvals = c(0.5, 0.5),
  layers = NULL, num = 4, amax = -250, amin = 70, border = 0.3,
  rstart = 0.02, flat = 1.2, show_legend = FALSE,
  show_label = TRUE, sorting = FALSE)
}
\arguments{
\item{data}{A four-column data frame. 1: population label (character); 2: region label (character); 3: genetic difference (numeric); 4: color code.}

\item{target}{A character value to pass name of the target population. Default is "target".}

\item{cenvals}{The coordinates of the center of the circle. Default is c(0.5, 0.5)}

\item{layers}{Numeric values of layers. This parameter will musk "num" if specified.}

\item{num}{An integer. Default is 4. Layer number.}

\item{amax}{Numeric. Maximal angle of radiation bars. Default is -250.}

\item{amin}{Numeric. Minimal angle of radiation bars. Default is 70.}

\item{border}{A numeric value of border. Default is 0.3.}

\item{rstart}{A numeric value of radius. Default is 0.02.}

\item{flat}{A numeric value. Flattening coefficient of radiation bars. Default is 1.2.}

\item{show_legend}{Logical. Whether to draw the legend of region information. Default is FALSE.}

\item{show_label}{Logical. Whether to print the population labels. Default is TRUE.}

\item{sorting}{Logical. Whether to sort population order according to their genetic difference. Default is FALSE.}
}
\description{
This function draws a radiation plot to visualize the genetic distance
}
