% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_plot.R
\name{sectorplot}
\alias{sectorplot}
\title{Sector plot}
\usage{
sectorplot(Q, ind, target = NULL, poporder = NULL, popgroup = NULL,
  ancescols = NULL, sort_order = FALSE, rmin = 2, rmax = 3.7,
  amin = -265, amax = 85, tar_ang1 = 0, tar_ang2 = 360,
  arrow = FALSE, show_legend = FALSE, ancesnames = NULL,
  prgap = 0.2, noline = FALSE)
}
\arguments{
\item{Q}{A numeric data frame of ancestry proportion (columns: ancestry componenyt; rows: individual). e.g., an output ".Q" file of the softaware ADMIXTURE.}

\item{ind}{A two-column data frame (1: population; 2: individual)."}

\item{target}{Character. The target populations to be plotted as a pie chart in the center of the circle figure. The population must be included in input "ind" and "Q" data frame.}

\item{poporder}{Character. The populations to be included in the figure, also the display order of the populations in the figure.}

\item{popgroup}{Character. The group information of populations.}

\item{ancescols}{The color code of each ancestry component in the figure.}

\item{sort_order}{A logical value to define whether to sort the order of the populations, which will be musked if "poporder" is specified.}

\item{rmin}{The radium of the inner ring. Default is 2.}

\item{rmax}{The radium of the outer ring. Default is 3.7.}

\item{amin}{The angle at which the ring is initiated. Defualt is -265.}

\item{amax}{The angle at which the ring is ended. Default is 85.}

\item{tar_ang1}{The start angle of the target leyout. Default is 0.}

\item{tar_ang2}{The end angle of the target leyout. Defualt is 360.}

\item{arrow}{Logical. Whether to draw the arrows to the target pies.}

\item{show_legend}{Logical. Whether to draw the legend of ancestry components.}

\item{ancesnames}{Character. To specify the names of ancestry components. If not specified, would be shown as "Ancestry_1" "Ancestry_2" and so on.}

\item{prgap}{A numeric value for gap length. Default is 0.2.}

\item{noline}{Logical. Whether to remove the black lines between populations. Default is FALSE.}
}
\description{
This function draws a sector plot to visualize ancestry proportion.
}
